+++ fxn/beta/metadata.py
@@ -27,7 +27,7 @@ def _validate_ort_inference_session (session: "onnxruntime.InferenceSession") ->
 
 class CoreMLInferenceMetadata (BaseModel):
     """
-    Metadata required to lower PyTorch models for inference on iOS, macOS, and visionOS with CoreML.
+    Metadata required to lower a PyTorch model for inference on iOS, macOS, and visionOS with CoreML.
     """
     kind: Literal["meta.inference.coreml"] = "meta.inference.coreml"
     model: Annotated[object, BeforeValidator(_validate_torch_module)] = Field(description="PyTorch module to apply metadata to.")
@@ -36,7 +36,7 @@ class CoreMLInferenceMetadata (BaseModel):
 
 class ONNXInferenceMetadata (BaseModel):
     """
-    Metadata required to lower PyTorch models for inference.
+    Metadata required to lower a PyTorch model for inference.
     """
     kind: Literal["meta.inference.onnx"] = "meta.inference.onnx"
     model: Annotated[object, BeforeValidator(_validate_torch_module)] = Field(description="PyTorch module to apply metadata to.")
@@ -45,17 +45,26 @@ class ONNXInferenceMetadata (BaseModel):
 
 class ONNXRuntimeInferenceSessionMetadata (BaseModel):
     """
-    Metadata required to lower ONNXRuntime inference sessions for inference.
+    Metadata required to lower an ONNXRuntime `InferenceSession` for inference.
     """
     kind: Literal["meta.inference.onnxruntime"] = "meta.inference.onnxruntime"
     session: Annotated[object, BeforeValidator(_validate_ort_inference_session)] = Field(description="ONNXRuntime inference session to apply metadata to.")
     model_path: Path = Field(description="ONNX model path. The model must exist at this path in the compiler sandbox.")
     model_config = ConfigDict(arbitrary_types_allowed=True, frozen=True)
 
-class GGUFInferenceMetadata (BaseModel): # INCOMPLETE
+class LlamaCppInferenceMetadata (BaseModel): # INCOMPLETE
     """
-    Metadata required to lower GGUF models for LLM inference.
+    Metadata required to lower a GGUF model for LLM inference.
     """
     kind: Literal["meta.inference.gguf"] = "meta.inference.gguf"
     model_path: Path = Field(description="GGUF model path. The model must exist at this path in the compiler sandbox.")
+    model_config = ConfigDict(arbitrary_types_allowed=True, frozen=True)
+
+class LiteRTInferenceMetadata (BaseModel):
+    """
+    Metadata required to lower PyTorch model for inference with LiteRT (fka TensorFlow Lite).
+    """
+    kind: Literal["meta.inference.litert"] = "meta.inference.litert"
+    model: Annotated[object, BeforeValidator(_validate_torch_module)] = Field(description="PyTorch module to apply metadata to.")
+    model_args: list[object] = Field(description="Positional inputs to the model.")
     model_config = ConfigDict(arbitrary_types_allowed=True, frozen=True)