+++ Changelog.md
@@ -1,5 +1,6 @@
 ## 0.0.48
-*INCOMPLETE*
++ Added `beta.LiteRTInferenceMetadata` class to lower PyTorch models for inference with LiteRT (fka TensorFlow Lite).
++ Refactored `beta.GGUFInferenceMetadata` class to `beta.LlamaCppInferenceMetadata`.
 
 ## 0.0.47
 + Fixed `Sandbox.upload_file` function failing to upload files when `from_path` is not an absolute path.
@@ -7,7 +8,7 @@
 
 ## 0.0.46
 + Added `metadata` parameter in `@compile` decorator to provide metadata to the compiler.
-+ Added `beta.CoreMLInferenceMetadata` class to lower PyTorch models to CoreML for inference on iOS, macOS, and visionOS.
++ Added `beta.CoreMLInferenceMetadata` class to lower PyTorch models for inference on iOS, macOS, and visionOS with CoreML.
 + Added `beta.ONNXInferenceMetadata` class to lower PyTorch models to ONNX for inference.
 + Added `beta.ONNXRuntimeInferenceSessionMetadata` class to lower ONNXRuntime inference sessions for inference.
 + Added `beta.GGUFInferenceMetadata` class to lower GGUF models for LLM inference.